# swagger.yaml
openapi: 3.0.3
info:
  title: New Relic SLM Handson API
  description: |
    New Relic Service Level Management (SLM) ハンズオン用のECサイトAPIです。
    
    このAPIは以下の機能を提供します：
    - 商品一覧・詳細の取得
    - ショッピングカート操作
    - 注文処理
    - SLMデモ用のエラー生成エンドポイント
    
    ## レスポンス形式
    すべてのAPIレスポンスは以下の形式で返されます：
    ```json
    {
      "success": true,
      "data": { /* 実際のデータ */ }
    }
    ```
    
    エラー時は以下の形式：
    ```json
    {
      "success": false,
      "error": {
        "code": "ERROR_CODE",
        "message": "エラーメッセージ"
      }
    }
    ```
  version: 1.0.0
  contact:
    name: NRUG-SRE
    url: https://github.com/NRUG-SRE/slm-handson
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:8080
    description: Local development server
  - url: http://api-server:8080
    description: Docker container (internal network)

paths:
  /health:
    get:
      summary: ヘルスチェック
      description: APIサーバーの稼働状況を確認するヘルスチェックエンドポイント
      tags:
        - Health
      responses:
        '200':
          description: サーバーが正常に稼働中
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "ok"
                  timestamp:
                    type: string
                    format: date-time
                    example: "2025-07-30T04:20:19Z"

  /api/products:
    get:
      summary: 商品一覧取得
      description: ECサイトの商品一覧を取得します。SLI測定の対象エンドポイントです。
      tags:
        - Products
      responses:
        '200':
          description: 商品一覧の取得に成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
        '500':
          description: サーバー内部エラー（SLMデモ用のランダムエラーを含む）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/products/{id}:
    get:
      summary: 商品詳細取得
      description: 指定されたIDの商品詳細情報を取得します
      tags:
        - Products
      parameters:
        - name: id
          in: path
          required: true
          description: 商品ID
          schema:
            type: string
            format: uuid
            example: "fa03a45e-8138-41b7-b2af-ed63881fb5f9"
      responses:
        '200':
          description: 商品詳細の取得に成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '404':
          description: 指定された商品が見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/cart:
    get:
      summary: カート内容取得
      description: 現在のカート内容を取得します。ユーザージャーニーの重要な部分です。
      tags:
        - Cart
      responses:
        '200':
          description: カート内容の取得に成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
        '500':
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/cart/items:
    post:
      summary: 商品をカートに追加
      description: 指定された商品をカートに追加します
      tags:
        - Cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - productId
                - quantity
              properties:
                productId:
                  type: string
                  format: uuid
                  description: 商品ID
                  example: "fa03a45e-8138-41b7-b2af-ed63881fb5f9"
                quantity:
                  type: integer
                  minimum: 1
                  description: 追加する数量
                  example: 2
      responses:
        '200':
          description: カートへの追加に成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
        '400':
          description: リクエストボディが不正
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 指定された商品が見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 在庫不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/cart/items/{id}:
    put:
      summary: カート内商品の数量変更
      description: カート内の指定商品の数量を変更します
      tags:
        - Cart
      parameters:
        - name: id
          in: path
          required: true
          description: カートアイテムID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - quantity
              properties:
                quantity:
                  type: integer
                  minimum: 0
                  description: 新しい数量（0で削除）
                  example: 3
      responses:
        '200':
          description: 数量変更に成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
        '400':
          description: リクエストボディが不正
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 指定されたカートアイテムが見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 在庫不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/orders:
    get:
      summary: 全注文一覧取得
      description: |
        システム内の全注文履歴を取得します。管理者用のエンドポイントです。
        ハンズオンでは注文履歴の確認に使用できます。
      tags:
        - Orders
      responses:
        '200':
          description: 注文一覧の取得に成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
        '500':
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: 注文作成
      description: |
        カート内容を基に注文を作成します。最も重要なビジネスKPIを測定するエンドポイントです。
        注文作成後、カートは空になります。
      tags:
        - Orders
      responses:
        '201':
          description: 注文作成に成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '422':
          description: カートが空、または在庫不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/orders/{id}:
    get:
      summary: 注文詳細取得
      description: 指定されたIDの注文詳細を取得します
      tags:
        - Orders
      parameters:
        - name: id
          in: path
          required: true
          description: 注文ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 注文詳細の取得に成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '404':
          description: 指定された注文が見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/error:
    get:
      summary: SLMデモ用エラー生成
      description: |
        SLM（Service Level Management）デモンストレーション用のエンドポイントです。
        環境変数 ERROR_RATE に基づいてランダムにエラーを発生させます。
        
        - ERROR_RATE=0.1 → 10%の確率でエラー
        - ERROR_RATE=0.5 → 50%の確率でエラー
      tags:
        - Demo
      responses:
        '200':
          description: 正常レスポンス
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: "This is a demo endpoint for SLM"
                      timestamp:
                        type: string
                        format: date-time
        '500':
          description: SLMデモ用の意図的なエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: 商品ID
          example: "fa03a45e-8138-41b7-b2af-ed63881fb5f9"
        name:
          type: string
          description: 商品名
          example: "ワイヤレスヘッドホン"
        description:
          type: string
          description: 商品説明
          example: "高音質なノイズキャンセリング機能付きワイヤレスヘッドホン"
        price:
          type: integer
          description: 価格（円）
          example: 25000
        imageUrl:
          type: string
          description: 商品画像URL
          example: "/images/headphones.jpg"
        stock:
          type: integer
          description: 在庫数
          example: 10
        createdAt:
          type: string
          format: date-time
          description: 作成日時
        updatedAt:
          type: string
          format: date-time
          description: 更新日時

    CartItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: カートアイテムID
        productId:
          type: string
          format: uuid
          description: 商品ID
        product:
          $ref: '#/components/schemas/Product'
        quantity:
          type: integer
          description: 数量
          example: 2
        subtotal:
          type: integer
          description: 小計（円）
          example: 50000
        addedAt:
          type: string
          format: date-time
          description: カート追加日時

    Cart:
      type: object
      properties:
        id:
          type: string
          description: カートID
          example: "default"
        items:
          type: array
          items:
            $ref: '#/components/schemas/CartItem'
        totalAmount:
          type: integer
          description: 合計金額（円）
          example: 50000
        itemCount:
          type: integer
          description: 商品点数
          example: 2
        updatedAt:
          type: string
          format: date-time
          description: 最終更新日時

    OrderItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: 注文アイテムID
        productId:
          type: string
          format: uuid
          description: 商品ID
        productName:
          type: string
          description: 商品名（注文時点）
          example: "ワイヤレスヘッドホン"
        price:
          type: integer
          description: 単価（注文時点）
          example: 25000
        quantity:
          type: integer
          description: 数量
          example: 2
        subtotal:
          type: integer
          description: 小計
          example: 50000

    Order:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: 注文ID
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        totalAmount:
          type: integer
          description: 注文合計金額（円）
          example: 50000
        status:
          type: string
          enum: ["pending", "processing", "completed", "cancelled"]
          description: 注文ステータス
          example: "completed"
        createdAt:
          type: string
          format: date-time
          description: 注文作成日時
        updatedAt:
          type: string
          format: date-time
          description: 最終更新日時

    ProductListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/Product'

    ProductResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Product'

    CartResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Cart'

    OrderResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Order'

    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              description: エラーコード
              enum:
                - "BAD_REQUEST"
                - "NOT_FOUND"
                - "UNPROCESSABLE_ENTITY"
                - "INTERNAL_SERVER_ERROR"
              example: "NOT_FOUND"
            message:
              type: string
              description: エラーメッセージ
              example: "Product not found"

tags:
  - name: Health
    description: ヘルスチェック関連
  - name: Products
    description: 商品管理
  - name: Cart
    description: ショッピングカート
  - name: Orders
    description: 注文管理
  - name: Demo
    description: SLMデモ用エンドポイント